Add prompts

+++ Client/src/app/shared/meals/meals.component.ts
@@ -68,13 +68,15 @@ export class MealsComponent implements OnInit {
     }
 
     onDeleteClick(item) {
-        this.dataService.deleteMeal(this.userId, item._id).subscribe(
-            data => {
-                this.meals = this.meals.filter(t => t._id !== item._id)
-                this.caloriesTrackingSubjectService.updated$.next()
-            },
-            error => this.sb.emitErrorSnackBar(error)
-        )
+        if (confirm(`Are you sure you want to delete ${item.name} meal`)) {
+            this.dataService.deleteMeal(this.userId, item._id).subscribe(
+                data => {
+                    this.meals = this.meals.filter(t => t._id !== item._id)
+                    this.caloriesTrackingSubjectService.updated$.next()
+                },
+                error => this.sb.emitErrorSnackBar(error)
+            )
+        }
     }
 
     onEditMealClick(item: Meal) {
